transition of cookie args to a dictionary
The following is an experiment on not commenting just about what this
specific patch does, but also why I did it, what I learned doing it,
etc.

Formerly, if the browser presented these cookies:

  name=Tom; age=36; occupation=SadMan

it would come back from .web.ck.decode (now called .web.ck.get,
though with no one using this API I don't think it matters) like so:

  (("name";"Tom");("age";"36");("occupation";"SadMan"))

This worked fine but it wasn't idiomatic. It was crude. At the time
I was just happy to get it together (it was one of my earliest
successful lines of code[1]). Here she is:

  // decode Cookie: headers into ((c1;v1);(c2;v2)...)
  ck.decode:{("=" vs) each "; " vs x}

I converted it slowly to this final form:

  ck.get:{pairs:("=" vs) each "; " vs x; (`$pairs[;0])!(pairs[;1])}

Note that we run a second pass over the data to basically build a
dictionary like so:

  `name`age`occupation!enlist "Tom" "36" "SadMan"

Sadly, I couldn't figure out a way to expand both "halves" of each
item of pairs, and then operate on them, in place.

I've actually had this problem in Q a few times, which makes me
wonder if I'm not missing a much bigger pattern:

Instead of applying a single function with a list of data (i.e.,
most of the adverbs as I understand them), apply a single piece
of data to a list of functions, unpacking as you go.

This is remarkably simple in Erlang:

  { [func1(Left),func2(Right)] <- List }.

Surely Q has a standard idiom to "explode" data - even Javascript
is getting this.

Anyway, I figured this out pretty quickly using an iterative approach
in the CLI that I've gotten better at. In fact, this is the only
way that works for me - to apply one more function at a time,
slowly massaging the data into a form that works.

It sounds clunky but it's much faster than writing big blocks of
code and then trying to fix them all at once.

I've seen APL programmers work this way on Youtube, and now I see why.