maintain "session" fk constraints
Next I had to clean up our session cookie support.

I transitioned to using symbols instead of strings as flags.

I also wrote what I feel is a pretty interesting attempt at
validating the session foreign key before inserting, so that
browsers with invalid or very old session keys don't cause an
foreign key exception.

First I just had to get the key out of the cookie store and into
a Q GUID[1]. That was a challenge, because I had something like:

  sess:$["G"$cookies`qwas;mksessid[]]

That was idiotic of course, because I don't think $[] supports
just two conditions - or does it? You can't test the truthiness
of a GUID anyway, so the first part was triggering 'type.

It should have been obvious to me but it was incredibly hard to
figure out what was going on.

I figured out I could test that the value was non-null with:

  null~"G"$cookies`qwas;

And I was on my way.

But of course another bug was lurking. Think if you can figure out
what it is.

Clue: Q is very serious about its foreign key constraints.

I ended up with this:

  sess:$[not (sess in (key sessions)`id) or null~sess:"G"$cookies `qwas;mksessid[];sess];

I think I could have used a select here instead, or maybe where by
itself?

[1] http://code.kx.com/wiki/Reference/Datatypes bottom of page